Respond to review feedback
This commit includes the following modifications:
  1. Apply the `java-library` plugin if neither `java` nor
     `java-library` have already been applied to the project.
  2. Don't recreate the classpath, just use it before setting it to
     empty.
  3. Remove any setting of `-sourcepath`.
  4. Add `inputs.property` for moduleName for caching.
  5. Use `findByName` to get the exact task we are looking for.

At this point, the plugin is faithfully executing the same changes
as the thrid step of the user guide. I will also add a step 4 to
the user guide which shows how to use the plugn.

Once this branch is reviewed, I will change the publication
configuration to publish to the Gradle Plugin Portal.